change logic
Delete edit_name actor to avoid having to edit a huge number of custom_attributes on possibly a lot of users.
This allows me to take advantage of the UserForm, users already signup won't see the custom attribute anymore (in case of active => inactive).
In case of inactive => active, users already signup will juste see the custom attribute again (if it was already present)z